+++ Sources/KeyboardKit/Layout/KeyboardLayout.swift
@@ -27,16 +27,16 @@ public class KeyboardLayout {
      Create a new layout with the provided `items`.
      
      - Parameters:
-       - items: The layout item rows to show in the keyboard.
+       - itemRows: The layout item rows to show in the keyboard.
     */
-    public init(items: KeyboardLayoutItemRows) {
-        self.items = items
+    public init(itemRows: KeyboardLayoutItemRows) {
+        self.itemRows = itemRows
     }
     
     /**
      The layout item rows to show in the keyboard.
      */
-    public let items: KeyboardLayoutItemRows
+    public let itemRows: KeyboardLayoutItemRows
     
     /**
      This `CGFloat` typealias makes it easier to see where a